+++ src/parser/companions/types/EldritchCannon2024.mjs
@@ -1,4 +1,4 @@
-import { logger, utils } from "../../../lib/_module.mjs";
+import { logger } from "../../../lib/_module.mjs";
 import { SUMMONS_ACTOR_STUB } from "./_data.mjs";
 import DDBCompanionMixin from "../DDBCompanionMixin.mjs";
 
@@ -137,6 +137,7 @@ export async function getEldritchCannons2024({
         },
       },
     },
+    items: [],
     prototypeToken: {
       name: "Eldritch Cannon",
       width: 0.5,
@@ -160,32 +161,31 @@ export async function getEldritchCannons2024({
 
   const actionData = extractEldritchCannonAbilities(raw);
 
+  let stub = foundry.utils.deepClone(cannonStub);
+  const manager = new DDBCompanionMixin(raw, { forceRulesVersion: "2024" }, { addMonsterEffects: true });
+  manager.npc = stub;
+
   for (const cannon of cannons) {
-    let stub = foundry.utils.deepClone(cannonStub);
-
-    stub.name = `${cannon.name}`;
-    stub.prototypeToken.name = `${cannon.name}`;
-
     const actionText = actionData.find((a) => cannon.name.includes(a.name))?.content ?? "";
     const description = `<p><em><strong>${cannon.name}.</strong></em> ${actionText}`;
-    const manager = new DDBCompanionMixin(description, { forceRulesVersion: "2024" }, { addMonsterEffects: true });
-    manager.npc = stub;
-    const features = await manager.getFeature(description, "special");
-    stub.items = features;
-    stub = await DDBCompanionMixin.addEnrichedImageData(stub);
-    const enriched = foundry.utils.getProperty(document, "flags.monsterMunch.enrichedImages");
 
-    results[`EldritchCannon${utils.idString(cannon.name)}2024`] = {
-      name: cannon.name,
-      version: enriched ? "2" : "1",
-      required: null,
-      isJB2A: false,
-      needsJB2A: false,
-      needsJB2APatreon: false,
-      folderName: "Eldritch Cannon (2024)",
-      data: stub,
-    };
+    const features = await manager.getFeature(description, "special");
+    stub.items.push(...features);
   }
+  stub = await DDBCompanionMixin.addEnrichedImageData(stub);
+  const enriched = foundry.utils.getProperty(document, "flags.monsterMunch.enrichedImages");
+
+  results[`EldritchCannon2024`] = {
+    name: "Eldritch Cannon",
+    version: enriched ? "2" : "1",
+    required: null,
+    isJB2A: false,
+    needsJB2A: false,
+    needsJB2APatreon: false,
+    folderName: "Eldritch Cannon (2024)",
+    data: stub,
+  };
+
 
   // console.warn("EldritchCannon result", results);
   logger.verbose("Eldritch Cannon results", results);