+++ src/parser/enrichers/class/artificer/EldritchCannon.mjs
@@ -19,14 +19,17 @@ export default class EldritchCannon extends DDBEnricherData {
   get activity() {
     return {
       id: "summonEldriComp1",
-      type: "summon",
       targetType: "creature",
       noTemplate: true,
-      profileKeys: [
-        { count: 1, name: "EldritchCannonFlamethrower2014" },
-        { count: 1, name: "EldritchCannonForceBallista2014" },
-        { count: 1, name: "EldritchCannonProtector2014" },
-      ],
+      profileKeys: this.is2014
+        ? [
+          { count: 1, name: "EldritchCannonFlamethrower2014" },
+          { count: 1, name: "EldritchCannonForceBallista2014" },
+          { count: 1, name: "EldritchCannonProtector2014" },
+        ]
+        : [
+          { count: 1, name: "EldritchCannon2024" },
+        ],
       summons: {
         match: {
           proficiency: false,
@@ -48,16 +51,27 @@ export default class EldritchCannon extends DDBEnricherData {
     };
   }
 
-  get addToDefaultAdditionalActivities() {
-    return true;
-  }
-
-  get useDefaultAdditionalActivities() {
-    return true;
-  }
 
   get additionalActivities() {
     return [
+      {
+        action: {
+          name: "Eldritch Cannon: Flamethrower",
+          type: "class",
+        },
+      },
+      {
+        action: {
+          name: "Eldritch Cannon: Force Ballista",
+          type: "class",
+        },
+      },
+      {
+        action: {
+          name: "Eldritch Cannon: Protector",
+          type: "class",
+        },
+      },
       {
         constructor: {
           name: "Summon With Spell Slot",