+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
-- 2014 subclass parsing fixes for: Way of the Sun Soul, Oath of Conquest, Oath of Redemption, Gloom Stalker.
+- 2014 subclass parsing fixes for: Way of the Sun Soul, Oath of Conquest, Oath of Redemption, Gloom Stalker, Monster Slayer.
+- Artificer 2024 Artillerist summon did not work
 - More Pugilist fixes
 - A number of small feature parsing improvements.