+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Next Up
 
-- 2014 subclass parsing fixes for: Way of the Sun Soul, Oath of Conquest, Oath of Redemption, Gloom Stalker, Monster Slayer, Divine Soul, Shadow Magic.
+# 6.6.54
+
+- 2014 subclass parsing fixes for: Way of the Sun Soul, Oath of Conquest, Oath of Redemption, Gloom Stalker, Monster Slayer, Divine Soul, Shadow Magic, Storm Sorcery.
 - Artificer 2024 Artillerist summon did not work
 - More Pugilist fixes
 - A number of small feature parsing improvements.